MAJ_summer_sprint (#705)
* MKP-1473/symfony 7.4 php 8.5 (#703)

* MKP-1473/symfony 7.4 php 8.5

* MKP-1473/add reference.php in gitignore

* MKP-1473/add reference.php in gitignore

* MKP-1473/fix baba's comments

---------


* MKP-1460: chore(ci): add GitHub Actions workflow for running tests (#707)

* MKP-1460: chore(ci): add GitHub Actions workflow for running tests

* MKP-1460: chore(ci): configure test environment for GitHub Actions

* MKP-1460: refactor(db): update database host configuration in TestDatabaseCloneTrait

* MKP-1460: chore(ci): setup Node.js and build frontend assets

* MKP-1460: feat(user): add getUsername method to User class

* MKP-1460: chore(ci): update GitHub Actions to use latest action versions

* MKP-1460: refactor(user): remove username field from user responses

* MKP-1445: SENTRY -  Clean-up & Monitoring (#708)

* MKP-1445: SENTRY -  Clean-up & Monitoring

* NEO-1445: fix tests

* NEO-1445: fix tests and INFO logger

* MKP-1445: Baba's requests

* MKP-1445: fix sentry MKP-8C error

* NEO-1445: add logs

* NEO-1445: fix tests

* NEO-1445: fix copilot

* NEO-1445: fix copilot

* NEO-1445: Baba's requests and add tests

---------


* MKP-1461: chore(ci): add CI GitHub Actions documentation (#710)

* MKP-1461: chore(ci): add CI GitHub Actions documentation

* MKP-1461: chore(ci): update Composer install command for CI workflow

* MKP-1480: minor and patches updating (#711)

* MKP-1480: minor and patches updating

* NEO-1480: reintroduce updated at field in log account connection

* NEO-1480: reintroduce updated at field in log account connection

* NEO-1480: reintroduce updated at field in log account connection

* NEO-1480: reintroduce updated at field in log account connection

---------


* MKP-1477/update pest + make lint (#712)

* MKP-1477/update pest + make lint

* MKP-1477/add updated_at field on log account connection table to avoid preprod login issue

---------


* MKP-1474/update phpcodesniffer (#713)


* MKP-1471/update api-platform/core (#714)


* MKP-1483/easy major updates (#716)


* MKP-1482/update tailwind (#717)

* MKP-1482/update tailwind first try

* MKP-1482/fix after review + detecte lint errors

* MKP-1482/fix yarn lint errors and warnings

* MKP-1482/fix sr-only issue

* MKP-1482/fix copilot's comments

* MKP-1482/fix menuaccount width

* MKP-1482/use @ instead absolute path

* MKP-1482/add explicit _error on empty catch

* MKP-1482/fix some code

* MKP-1482/don't use absolute path for src and template in main.css

---------


* MKP-1481/update vite (#718)


* fix veronique's comments

* fix veronique's comments

---------

+++ src/Service/Channel/ChannelOptionSynchronizer.php
@@ -48,6 +48,9 @@ class ChannelOptionSynchronizer
         foreach ($optionsArray as $key => $value) {
             if (!\array_key_exists($key, $channelOptions)) {
                 $channelOption = $this->entityManager->getRepository(ChannelOption::class)->findOneBy(['name' => $key, 'channel' => $channel]);
+                if ($channelOption === null) {
+                    continue;
+                }
                 $this->entityManager->remove($channelOption);
             }
         }
